NP-7222 Move each panel to its own component
